requirements.txt
bleak
pystray
Pillow
keyring
pyautogui
plyer
pywin32
